UIMVC refractor P3 Controller (Xander's gonna document)
