+++ UI-MVC/Controllers/PlatformController.cs
@@ -1,11 +1,10 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using BL;
 using Domain.Identity;
-using Domain.Projects;
-using Domain.UserInput;
 using Domain.Users;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
@@ -18,16 +17,12 @@ namespace UIMVC.Controllers
     public class PlatformController : Controller
     {
         private readonly PlatformManager _platformMgr;
-        private readonly ProjectManager _projectMgr;
-        private readonly IdeationQuestionManager _iqMgr;
         private readonly UserManager<UimvcUser> _userManager;
         private readonly RoleService _roleService;
 
         public PlatformController(UserManager<UimvcUser> userManager, RoleService service)
         {
             _platformMgr = new PlatformManager();
-            _projectMgr = new ProjectManager();
-            _iqMgr = new IdeationQuestionManager();
             _userManager = userManager;
             _roleService = service;
         }
@@ -140,123 +135,178 @@ namespace UIMVC.Controllers
         }
 
         #endregion
-
-        #region Project
-
-        [HttpGet]
-        public IActionResult CollectProject(int id)
-        {
-            Project project = _projectMgr.GetProject(id, true);
-
-            if (project.Visible && project.Id != 0 || _roleService.IsAdmin(HttpContext.User).Result)
-            {
-                List<Phase> phases = (List<Phase>) _projectMgr.GetAllPhases(id);
-
-                foreach (Phase phase in phases)
-                {
-                    if (phase.Id == project.CurrentPhase.Id)
-                    {
-                        project.CurrentPhase = phase;
-                    }
-                }
-
-                phases.Remove(project.CurrentPhase);
-                ViewData["Phases"] = phases;
-
-                return View(project);
-            }
-
-            if (!project.Visible)
-            {
-                return RedirectToAction("Index", "Platform", new
-                {
-                    id = project.Platform.Id, message = "Project is niet ingesteld voor het algemene publiek!"
-                });
-            }
-
-            
-            return RedirectToAction("HandleErrorCode", "Errors", 
-                new { statuscode = 404, path="/Platform/CollectProject/" + id  });
-        }
-
-        #endregion
-
-        #region Ideation
-
-        public IActionResult CollectIdeation(int id)
-        {
-            Ideation ideation = _projectMgr.ModuleMan.GetIdeation(id, true);
-
-            return View(ideation);
-        }
-
-        public IActionResult CollectIdeationThread(int id, string message)
-        {
-            IdeationQuestion iq = _iqMgr.GetQuestion(id, true);
-
-            ViewData["Message"] = message;
-            ViewData["IdeationQuestion"] = iq;
-
-            return View(iq);
-        }
         
-        #endregion
-
-        #region Idea
-
-        [Authorize]
-        public IActionResult AddVote(int idea, string user, int thread)
+        #region AddPlatform
+        [HttpGet]
+        [Authorize(Roles = "SuperAdmin")]
+        public IActionResult AddPlatform()
         {
-            if (_iqMgr.MakeVote(idea, user))
-            {
-                return RedirectToAction("CollectIdeationThread", "Platform", routeValues: new
-                    { id = thread, message = "Stem gelukt, dankjewel!" });
-            }
-
-            return RedirectToAction("CollectIdeationThread", "Platform", routeValues: new
-                {id = thread, message = "Al gestemd op dit idee!"});
+            ViewData["platforms"] = _platformMgr.ReadAllPlatforms();
+            return View();
         }
 
-        [Authorize]
-        public IActionResult AddReport(int idea, string flagger, int thread)
+        [HttpPost]
+        [Authorize(Roles = "SuperAdmin")]
+        public async Task<IActionResult> AddPlatform(AddPlatformModel cpm)
         {
-            Idea ToReport = _iqMgr.GetIdea(idea, false);
-
-            Report report = new Report()
+            if (cpm == null)
             {
-                Idea = ToReport,
-                Flagger = new UimvcUser() {Id = flagger},
-                Reportee = new UimvcUser() {Id = ToReport.User.Id},
-                Status = ReportStatus.StatusNotViewed
+                return BadRequest("Platform cannot be null");
+            }
+
+            Platform platform = new Platform()
+            {
+                Name = cpm.Name,
+                Url = cpm.Url,
+                Owners = new List<UimvcUser>(),
+                Users = new List<UimvcUser>()
             };
 
-            if (!Request.Form["Reason"].ToString().Equals(""))
+            using (var memoryStream = new MemoryStream())
             {
-                Report alreadyReport = _iqMgr.GetAllReportsByIdea(idea).FirstOrDefault(
-                    r => r.Flagger.Id.Equals(flagger));
-
-                if (alreadyReport == null)
-                {
-                    report.Reason = Request.Form["Reason"].ToString();
-                    _iqMgr.MakeReport(report);
-
-                    ToReport.Reported = true;
-                    _iqMgr.EditIdea(ToReport);
-
-                    return RedirectToAction("CollectIdeationThread", "Platform", routeValues: new
-                        {id = thread, message = "Idee geraporteerd!"});
-                }
-
-                return RedirectToAction("CollectIdeationThread", "Platform", routeValues: new
-                    {id = thread, message = "Je oude rapport is nog in de behandeling"});
+                await cpm.IconImage.CopyToAsync(memoryStream);
+                platform.IconImage = memoryStream.ToArray();
             }
 
-            return RedirectToAction("CollectIdeationThread", "Platform", routeValues: new
-                {id = thread, message = "Je hebt geen reden opgegeven voor je rapport!"});
+            using (var memoryStream = new MemoryStream())
+            {
+                await cpm.CarouselImage.CopyToAsync(memoryStream);
+                platform.CarouselImage = memoryStream.ToArray();
+            }
+
+            using (var memoryStream = new MemoryStream())
+            {
+                await cpm.FrontPageImage.CopyToAsync(memoryStream);
+                platform.FrontPageImage = memoryStream.ToArray();
+            }
+
+            var newPlatform = _platformMgr.MakePlatform(platform);
+
+            return RedirectToAction("Index", "Platform", new {Id = newPlatform.Id} );
         }
-        
+
+        [HttpPost]
+        [Authorize(Roles = "Admin, SuperAdmin")]
+        public async Task<IActionResult> AssignUserToPlatform(AssignUserModel aum)
+        {
+            if (aum == null) return BadRequest("Cannot be null");
+            if (User.IsInRole(Role.Admin.ToString("G")) &&
+                (await _userManager.GetUserAsync(User)).PlatformDetails != aum.PlatformId)
+                return BadRequest("You are no admin of this platform");
+
+            UimvcUser user = await _userManager.FindByEmailAsync(aum.UserMail);
+            if (user == null) return BadRequest("Wrong user mail");
+            user.PlatformDetails = aum.PlatformId;
+
+            if (aum.Role == 0) aum.Role = AssignUserRole.MODERATOR;
+            _userManager.AddToRoleAsync(user, Enum.GetName(typeof(AssignUserRole), aum.Role));
+
+            await _userManager.UpdateAsync(user);
+
+            return RedirectToAction("ChangePlatform", "Platform", new {Id = aum.PlatformId} );
+        }
+
+        [HttpPost]
+        [Authorize(Roles = "Admin, SuperAdmin")]
+        public async Task<IActionResult> RemoveUserFromPlatform(AssignUserModel aum)
+        {
+            if (aum == null)
+            {
+                return BadRequest("Cannot be null");
+            }
+
+
+            UimvcUser user = await _userManager.FindByEmailAsync(aum.UserMail);
+            if (user == null) return BadRequest("Wrong user mail");
+            user.PlatformDetails = 0;
+            if (await _userManager.IsInRoleAsync(user, "Admin") || await _userManager.IsInRoleAsync(user, "Moderator"))
+            {
+                _userManager.RemoveFromRolesAsync(user, new[] {"Moderator", "Admin"});
+            }
+
+            await _userManager.UpdateAsync(user);
+
+            return RedirectToAction("ChangePlatform", "Platform", new {Id = aum.PlatformId} );
+        }
+
         #endregion
+        
+        #region UIMVCUser
+        [HttpGet]
+        [Authorize(Roles = "Moderator, Admin, SuperAdmin")]
+        public async Task<IActionResult> CollectAllUsers(string sortOrder, string searchString)
+        {
 
+            ViewData["CurrentFilter"] = searchString;
+            var users = (IEnumerable<UimvcUser>)_userManager.Users;
 
+            if (!String.IsNullOrEmpty(searchString))
+            {
+                users = users.Where(u => u.Name.ToUpper().Contains(searchString.ToUpper()));
+            }
+            if (!User.IsInRole(Role.SuperAdmin.ToString("G")))
+            {
+                UimvcUser user = await _userManager.GetUserAsync(User);
+                users = users.Where(u => u.PlatformDetails == user.PlatformDetails);
+            }
+
+            switch (sortOrder)
+            {
+                case "platform":
+                    users = users.OrderBy(u => u.PlatformDetails); break;
+                case "name":
+                    users = users.OrderBy(u => u.Name); break;
+                case "birthday":
+                    users = users.OrderBy(u => u.DateOfBirth); break;
+                default:
+                    users = users.OrderBy(u => u.Id); break;
+            }
+            return View(users);
+        }
+
+        [HttpGet]
+        [Authorize(Roles = "Moderator, Admin, SuperAdmin")]
+        public async Task<IActionResult> ToggleBanUser(string userId)
+        {
+            UimvcUser userFound = await _userManager.FindByIdAsync(userId);
+
+            if (userFound == null) return RedirectToAction("CollectAllUsers");
+            if (await _roleService.IsSameRoleOrHigher(HttpContext.User, userFound)) return RedirectToAction("CollectAllUsers");
+
+            userFound.Banned = !userFound.Banned;
+            _userManager.SetLockoutEnabledAsync(userFound, userFound.Banned);
+            if (userFound.Banned)
+            {
+                _userManager.SetLockoutEndDateAsync(userFound, DateTime.MaxValue);
+            }
+            var result = await _userManager.UpdateAsync(userFound);
+
+            return RedirectToAction("CollectAllUsers");
+        }
+
+        [Authorize(Roles = "Admin, SuperAdmin")]
+        public async Task<IActionResult> SetRole(AssignRoleModel arm, string userId)
+        {
+            var user = await _userManager.FindByIdAsync(userId);
+            string roletext = Request.Form["Role"];
+            Object roleParse = null;
+            if (!Enum.TryParse(typeof(Role), roletext, out roleParse)) return RedirectToAction("CollectAllUsers", 
+                "Platform");
+            var role = (Role) roleParse;
+            
+            if (!await _roleService.IsSameRoleOrLower(User, role))
+            {
+                if (await _userManager.IsInRoleAsync(user, roletext))
+                {
+                    await _userManager.RemoveFromRoleAsync(user, roletext);
+                }
+                else
+                {
+                    _roleService.AssignToRole(user, role);
+                }
+            }
+            return RedirectToAction("CollectAllUsers", "Platform");
+        }
+        #endregion
     }
 }
