+++ UI-MVC/Controllers/ProjectController.cs
@@ -8,9 +8,9 @@ using Domain.Projects;
 using Domain.Users;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using UIMVC.Models;
+using UIMVC.Services;
 
 namespace UIMVC.Controllers
 {
@@ -18,21 +18,54 @@ namespace UIMVC.Controllers
     {
         private ProjectManager _projManager;
         private ModuleManager _modManager;
-        private readonly UserManager<UimvcUser> _userManager;
+        private readonly RoleService _roleService;
 
-
-         public ProjectController(UserManager<UimvcUser> userManager)
+        public ProjectController(RoleService service)
         {
             _modManager = new ModuleManager();
             _projManager = new ProjectManager();
-            _userManager = userManager;
+            _roleService = service;
         }
 
          #region Project
+         [HttpGet]
+         public IActionResult CollectProject(int id)
+         {
+             Project project = _projManager.GetProject(id, true);
 
+             if (project.Visible && project.Id != 0 || _roleService.IsAdmin(HttpContext.User).Result)
+             {
+                 List<Phase> phases = (List<Phase>) _projManager.GetAllPhases(id);
+
+                 foreach (Phase phase in phases)
+                 {
+                     if (phase.Id == project.CurrentPhase.Id)
+                     {
+                         project.CurrentPhase = phase;
+                     }
+                 }
+
+                 phases.Remove(project.CurrentPhase);
+                 ViewData["Phases"] = phases;
+
+                 return View(project);
+             }
+
+             if (!project.Visible)
+             {
+                 return RedirectToAction("Index", "Platform", new
+                 {
+                     id = project.Platform.Id, message = "Project is niet ingesteld voor het algemene publiek!"
+                 });
+             }
+
+            
+             return RedirectToAction("HandleErrorCode", "Errors", 
+                 new { statuscode = 404, path="/Platform/CollectProject/" + id  });
+         }
+         
          #region Add
-
-        [Authorize(Roles ="Admin, SuperAdmin")]
+         [Authorize(Roles ="Admin, SuperAdmin")]
         [HttpGet]
         public IActionResult AddProject(int platform)
         {
@@ -96,11 +129,10 @@ namespace UIMVC.Controllers
                 _projManager.MakeProjectImage(memoryStream.ToArray(), projectId);
             }
 
-            return RedirectToAction("CollectProject", "Platform",
+            return RedirectToAction("CollectProject", "Project",
                 new {id = projectId});
         }
          #endregion
-
 
          #region ChangeProject
 
@@ -137,7 +169,7 @@ namespace UIMVC.Controllers
             }
 
             _projManager.EditProject(updateProj);
-            return RedirectToAction("CollectProject", "Platform", new {id = updateProj.Id});
+            return RedirectToAction("CollectProject", "Project", new {id = updateProj.Id});
         }
 
          #endregion
@@ -192,12 +224,9 @@ namespace UIMVC.Controllers
 
          #endregion
 
-         #region phase
-
+         #region Phase
          #region AddPhase
-
          [Authorize(Roles ="Admin, SuperAdmin")]
-
          [HttpGet]
         public IActionResult AddPhase(int projectId)
         {
@@ -205,8 +234,7 @@ namespace UIMVC.Controllers
 
              return View();
         }
-
-
+        
          [Authorize(Roles ="Admin, SuperAdmin")]
         [HttpPost]
         public IActionResult AddPhase(PhaseViewModel pm, int projectId)
@@ -226,15 +254,13 @@ namespace UIMVC.Controllers
 
              _projManager.MakePhase(p);
 
-             return RedirectToAction("CollectProject", "Platform", new {id = projectId});
+             return RedirectToAction("CollectProject", "Project", new {id = projectId});
         }
 
          #endregion
-
-
+         
          #region ChangePhase
-
-        [Authorize(Roles ="Admin, SuperAdmin")]
+         [Authorize(Roles ="Admin, SuperAdmin")]
         [HttpGet]
         public IActionResult ChangePhase(int phaseId)
         {
@@ -262,13 +288,11 @@ namespace UIMVC.Controllers
             updatePhase.EndDate = pm.EndDate;
 
              _projManager.EditPhase(updatePhase);
-            return RedirectToAction("CollectProject", "Platform", new {id = updatePhase.Project.Id});
+            return RedirectToAction("CollectProject", "Project", new {id = updatePhase.Project.Id});
         }
 
          #endregion
-
-
-
+         
          [Authorize(Roles ="Admin, SuperAdmin")]
         [HttpGet]
         public IActionResult SetCurrentPhase(int projectId, int phaseId)
@@ -281,7 +305,7 @@ namespace UIMVC.Controllers
 
              _projManager.EditProject(p);
 
-             return RedirectToAction("CollectProject", "Platform", new {id = projectId});
+             return RedirectToAction("CollectProject", "Project", new {id = projectId});
         }
 
          #region DestroyPhase
@@ -293,11 +317,31 @@ namespace UIMVC.Controllers
         {
             _projManager.RemovePhase(phaseId);
 
-             return RedirectToAction("CollectProject", "Platform", new {id = projectId});
+             return RedirectToAction("CollectProject", "Project", new {id = projectId});
         }
 
          #endregion
 
+         #endregion
+         
+         #region Tags
+         [Authorize(Roles = "Admin, SuperAdmin")]
+         public IActionResult AddTag(int ideation, bool questionnaire = false)
+         {
+             string tag = Request.Form["GetMeATag"].ToString();
+
+             if (tag == null)
+             {
+                 return BadRequest("Tag can't be null");
+             }
+            
+             _modManager.MakeTag(tag, ideation, questionnaire);
+
+             if (questionnaire)
+                 return RedirectToAction("ChangeQuestionnaire", "Questionnaire", new {questionnaireId = ideation});
+
+             return RedirectToAction("CollectIdeation", "Ideation", new {Id = ideation});
+         }
          #endregion
     }
 }
