+++ UI-MVC/UIMVC.csproj
@@ -69,4 +69,22 @@
     <Folder Include="Areas\Identity\Data" />
   </ItemGroup>
 
+  <ItemGroup>
+    <_ContentIncludedByDefault Remove="Views\Moderation\ModerationTemplates\AssignRolePartial.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Moderation\ModerationTemplates\SidebarView.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\Thread\FieldDisplay\ClosedField.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\Thread\FieldDisplay\Field.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\Thread\FieldDisplay\ImageField.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\Thread\FieldDisplay\MapField.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\Thread\FieldDisplay\VideoField.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\Thread\_AddIdea.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\Thread\_ChangeIdea.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\Thread\_CollectIdea.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\Thread\_IdeaFields.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\DisplayTemplates\Ideation.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\DisplayTemplates\Phase.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\DisplayTemplates\Project.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Platform\DisplayTemplates\Questionnaire.cshtml" />
+  </ItemGroup>
+
 </Project>
