+++ UI-MVC/Views/Ideation/AddCentralQuestion.cshtml
@@ -4,7 +4,7 @@
     <div class="row">
         <div class="col-md-10">
             <h1>Een nieuwe centrale vraag stellen.</h1>
-            <form asp-controller="Moderation" asp-action="AddCentralQuestion" method="post" asp-route-ciqm="@Model"
+            <form asp-controller="Ideation" asp-action="AddCentralQuestion" method="post" asp-route-ciqm="@Model"
                   asp-route-ideation="@ViewData["Ideation"]">
                 <div class="form-group">
                     <label>IdeationID</label>
