+++ UI-MVC/Views/Ideation/AddIdeation.cshtml
@@ -8,7 +8,7 @@
         <div class="col-md-10">
             <h1>Een Ideation aanmaken.</h1>
 
-            <form asp-controller="Moderation" asp-action="AddIdeation" method="post" asp-route-cim="@Model"
+            <form asp-controller="Ideation" asp-action="AddIdeation" method="post" asp-route-cim="@Model"
                   asp-route-project="@ViewData["Project"]" asp-route-user="@UserManager.GetUserId(User)">
                 <div class="form-group">
                     <label>ProjectID</label>
