+++ UI-MVC/Views/Ideation/ChangeIdeation.cshtml
@@ -9,7 +9,7 @@
     <div class="row">
         <div class="col-md-10">
             <h1>Een Ideation Aanpassen.</h1>
-            <form asp-controller="Moderation" asp-action="ConfirmChangeIdeation" method="post" 
+            <form asp-controller="Ideation" asp-action="ConfirmChangeIdeation" method="post" 
                   asp-route-ideation="@ViewData["Ideation"]" asp-route-aim="@Model">
                 <div class="form-group">
                     <label>ProjectID</label>
