+++ UI-MVC/Views/Ideation/CollectIdeationThread.cshtml
@@ -24,7 +24,7 @@
                 <p>Lees meer: <a href="">@Model.SiteUrl</a></p>
             }
             <br/>
-            <a class="btn btn-primary" asp-controller="Platform" asp-action="CollectIdeation"
+            <a class="btn btn-primary" asp-controller="Ideation" asp-action="CollectIdeation"
                asp-route-id="@Model.Ideation.Id">Terug Naar Ideation</a>
         </div>
     </div>
