+++ UI-MVC/Views/Ideation/FieldDisplay/ClosedField.cshtml
