+++ UI-MVC/Views/Ideation/FieldDisplay/ImageField.cshtml
