+++ UI-MVC/Views/Platform/AddPlatform.cshtml
@@ -55,7 +55,7 @@
      aria-labelledby="makePlatformModalLabel" aria-hidden="true">
     <div class="modal-dialog modal-dialog-centered" role="document">
         <div class="modal-content">
-            <form asp-controller="Moderation" asp-action="AddPlatform" method="post" enctype="multipart/form-data">
+            <form asp-controller="Platform" asp-action="AddPlatform" method="post" enctype="multipart/form-data">
                 <div class="modal-body m-auto">
                     <div class="form-group">
                         <label asp-for="Name">Naam van het platform</label>
