+++ UI-MVC/Views/Platform/Moderation/ModerationTemplates/AssignRolePartial.cshtml
@@ -1,7 +1,7 @@
 @using Domain.Users
 @model AssignRoleModel
 
-<form class="row" asp-action="SetRole" asp-controller="Moderation" asp-route-arm="@Model" asp-route-userId="@Model.UserId">
+<form class="row" asp-action="SetRole" asp-controller="Platform" asp-route-arm="@Model" asp-route-userId="@Model.UserId">
     <div class="col-10 m-0">
             @Html.DropDownListFor(model => model.Role, new SelectList(Enum.GetValues(typeof(Role))), "Selecteer een rol", new {@class = "form-control"})
