+++ UI-MVC/Views/Project/CollectProject.cshtml
@@ -174,7 +174,7 @@
 {
     <section class="container">
         <div class="row justify-content-around">
-            <a class="btn btn-primary" asp-controller="Moderation" asp-action="AddIdeation"
+            <a class="btn btn-primary" asp-controller="Ideation" asp-action="AddIdeation"
                asp-route-project="@Model.Id">
                 Maak een nieuwe ideation
             </a>
