+++ UI-MVC/Views/Project/Partials/_ProjectPhaseDetailPartial.cshtml
