--- UI-MVC/Views/Shared/_AddQuestionnaireQuestionPartial.cshtml.cs
@@ -1,11 +1,0 @@
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace UIMVC.Views.Shared
-{
-    public class IndexModel : PageModel
-    {
-        public void OnGet()
-        {
-        }
-    }
-}
