--- UI-MVC/Views/Shared/_Layout.cshtml.cs
@@ -1,16 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace UIMVC.Views.Shared
-{
-    public class _LayoutModel : PageModel
-    {
-        public void OnGet()
-        {
-        }
-    }
-}
